sqlite: implement size in .get_info method
The type of the size is changed from int32_t to int64_t.
An int32_t can only hold sizes up to 2GB which is not
enough for a database. A int64_t is more than enough to
hold a size.

+++ rbh-info/rbh-info.c
@@ -312,7 +312,7 @@ _get_collection_size(const struct rbh_value *value)
 
     buffer = _buffer;
 
-    size_printer(buffer, sizeof(_buffer), value->int32);
+    size_printer(buffer, sizeof(_buffer), value->int64);
 
     printf("%s\n", buffer);
 }
